move DM shape reading function from util to hicat simulator
