+++ pastis/util_pastis.py
@@ -236,21 +236,6 @@ def apply_mode_to_luvoir(pmode, luvoir):
     return planes['seg_mirror'], psf
 
 
-def read_continuous_dm_maps_hicat(path_to_dm_maps):
-    """
-    Read Boston DM maps from disk and return as one list per DM.
-    :param path_to_dm_maps: string, absolute path to folder containing DM maps to load
-    :return: DM1 actuator map array, DM2 actuator map array; in m
-    """
-
-    surfaces = []
-    for dmnum in [1, 2]:
-        actuators_2d = fits.getdata(os.path.join(path_to_dm_maps, f'dm{dmnum}_command_2d_noflat.fits'))
-        surfaces.append(actuators_2d)
-
-    return surfaces[0], surfaces[1]
-
-
 def segment_pairs_all(nseg):
     """
     Return a generator with all possible segment pairs, including repeating ones.
